requirements.txt
# Core
streamlit==1.32.0
langchain==0.3.0
langchain-community==0.3.0

# Embeddings - CPU only
fastembed==0.5.1
faiss-cpu==1.8.0

# PDF Processing
pypdf==5.1.0
pdfplumber==0.11.2

# For tiny models (cloud deployment)
transformers==4.42.0
torch==2.3.0
accelerate==0.31.0
sentencepiece==0.2.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
numpy==1.26.4
pandas==2.0.3
plotly==5.22.0 